+++ contracts/token/ERC20/facets/ERC20MetadataFacet.sol
@@ -12,20 +12,14 @@ import {ForwarderRegistryContextBase} from "./../../../metatx/base/ForwarderRegi
 /// @dev This contract is to be used as a diamond facet (see ERC2535 Diamond Standard https://eips.ethereum.org/EIPS/eip-2535).
 /// @dev Note: This facet depends on {ProxyAdminFacet}, {ContractOwnershipFacet} and {InterfaceDetectionFacet}.
 contract ERC20MetadataFacet is ERC20MetadataBase, ForwarderRegistryContextBase {
-    using ERC20MetadataStorage for ERC20MetadataStorage.Layout;
     using ProxyAdminStorage for ProxyAdminStorage.Layout;
 
     constructor(IForwarderRegistry forwarderRegistry) ForwarderRegistryContextBase(forwarderRegistry) {}
 
-    /// @notice Initializes the storage with an initial token URI.
-    /// @notice Sets the proxy initialization phase to `1`.
     /// @notice Marks the following ERC165 interface(s) as supported: ERC20Metadata.
-    /// @dev Reverts if the sender is not the proxy admin.
-    /// @dev Reverts if the proxy initialization phase is set to `1` or above.
-    /// @param uri The token URI.
-    function initERC20MetadataStorage(string calldata uri) external {
+    function initERC20MetadataStorage() external {
         ProxyAdminStorage.layout().enforceIsProxyAdmin(_msgSender());
-        ERC20MetadataStorage.layout().proxyInit(uri);
+        ERC20MetadataStorage.init();
     }
 
     /// @inheritdoc ForwarderRegistryContextBase